Arrumar tudo ;-;

+++ JulhoCamisas/Modelos/Acabamento.cs
@@ -5,7 +5,7 @@ namespace JulhoCamisas.Modelos;
 public class Acabamento : Registro{
   
   [BsonId]
-  public Int64 Id { get; set; }
+  public int Id { get; set; }
 
   public string Quantidade { get; set; }
   public string Tipo { get; set; }
